Post Media 컴포넌트 책임을 분리한다
